completed simplified tanakh api with dynamic api routing
