+++ app/api/tanakh/amos.ts
