+++ app/api/tanakh/chronicles2.ts
