+++ app/api/tanakh/daniel.ts
