+++ app/api/tanakh/deuteronomy.ts
