+++ app/api/tanakh/ecclesiastes.ts
