+++ app/api/tanakh/exodus.ts
