+++ app/api/tanakh/genesis.ts
