+++ app/api/tanakh/hosea.ts
