+++ app/api/tanakh/job.ts
