+++ app/api/tanakh/kings1.ts
