+++ app/api/tanakh/kings2.ts
