+++ app/api/tanakh/malachi.ts
