+++ app/api/tanakh/obadiah.ts
