+++ app/api/tanakh/ruth.ts
