+++ app/api/tanakh/samuel1.ts
