+++ app/api/tanakh/samuel2.ts
